refactor(dubbo): update code structure in triple mod
