+++ dubbo/src/triple/mod.rs
@@ -17,32 +17,9 @@
 
 pub mod client;
 pub mod codec;
-pub mod invocation;
+pub mod compression;
+pub mod consts;
+pub mod decode;
+pub mod encode;
 pub mod server;
 pub mod transport;
-
-// use http_body::Body;
-// use std::future::Future;
-// use std::pin::Pin;
-
-// pub(crate) type Error = Box<dyn std::error::Error + Send + Sync>;
-
-// pub type BoxBody = http_body::combinators::UnsyncBoxBody<bytes::Bytes, self::status::Status>;
-// pub type BoxFuture<T, E> = self::Pin<Box<dyn self::Future<Output = Result<T, E>> + Send + 'static>>;
-
-// pub fn empty_body() -> BoxBody {
-//     http_body::Empty::new()
-//         .map_err(|err| match err {})
-//         .boxed_unsync()
-// }
-
-// pub(crate) fn boxed<B>(body: B) -> BoxBody
-// where
-//     B: http_body::Body<Data = bytes::Bytes> + Send + 'static,
-//     B::Error: Into<self::Error>,
-// {
-//     body.map_err(|err| {
-//         self::status::Status::new(self::status::Code::Internal, format!("{:?}", err.into()))
-//     })
-//     .boxed_unsync()
-// }
