+++ dubbo/src/triple/server/mod.rs
@@ -15,12 +15,7 @@
  * limitations under the License.
  */
 
-pub mod compression;
-pub mod consts;
-pub mod decode;
-pub mod encode;
-pub mod server;
 pub mod service;
+pub mod triple;
 
-pub use decode::Decoding;
-pub use encode::{encode, encode_server};
+pub use triple::TripleServer;
