+++ dubbo/src/triple/server/service.rs
@@ -15,10 +15,11 @@
  * limitations under the License.
  */
 
-use super::decode::Decoding;
-use crate::triple::invocation::{Request, Response};
 use futures_util::{Future, Stream};
 use tower_service::Service;
+
+use crate::invocation::{Request, Response};
+use crate::triple::decode::Decoding;
 
 pub trait StreamingSvc<R> {
     type Response;
